v1.5.3 * LTE client type and event type/subtype for Wireless State Response added * Bugfix Wireless interface ID is now displayed correctly * Bugfix Update error for already-present interface IDs in MIB::updateWirelessState() is fixed * Simulator is now able to send unsolicited WIRELESS_STATE_RESPONSE

+++ trunk/openair3/SCOREF-MGMT/Software/src/packets/mgmt_gn_packet_wireless_state_response.cpp
@@ -66,7 +66,7 @@ bool GeonetWirelessStateResponseEventPacket::parse(const vector<unsigned char>& 
 
 	// Parse interface count first
 	u_int8_t interfaceCount = packetBuffer.data()[sizeof(MessageHeader)];
-	logger.info("Number of interfaces is " + boost::lexical_cast<string>(interfaceCount));
+	logger.info("Number of interfaces is " + boost::lexical_cast<string>((int)interfaceCount));
 
 	// Then traverse the buffer to get the state for every interface...
 	u_int16_t itemIndex = sizeof(WirelessStateResponseMessage);
